api/requirements.txt
flask
redis
pyjwt
python-dotenv
requests